Demonstration of insert_all failure
`insert_all` is failing because it regenerates the primary key (`id` in
this case) instead of updating an existing row.

The `id` column in referred to in the generated SQL statement and by
using the `DEFAULT` keyword in the `VALUES` clause of the insert
statement.

Run this test like this:

	MIX_ENV=pg mix test integration_test/cases/repo.exs

The output is:

	  1) test upsert via insert_all on conflict replace_all and conflict_target is unique index (Ecto.Integration.RepoTest)
	     integration_test/cases/repo.exs:1545
	     ** (UndefinedFunctionError) function nil.title/0 is undefined or private
	     stacktrace:
	       nil.title()
	       integration_test/cases/repo.exs:1565: (test)

The error is because oen of the original records can no longer be found
by its original `id`.